deploy/requirements.txt
aws-cdk.cdk
aws-cdk.aws_ec2
aws-cdk.aws_ecs
aws-cdk.aws_ssm
aws-cdk.aws_secretsmanager
aws-cdk.aws_ecs_patterns
aws-cdk.aws_iam
aws-cdk.aws_servicediscovery
aws-cdk.aws_logs
aws_cdk.aws_kinesis
aws_cdk.aws_kinesisfirehose
aws_cdk.aws_s3
